Move tests from condegen/src/lib.rs to codegen/tests directory.
